refresh cards after close button clicked
