+++ CSETWebNg/src/app/initial/new-assessment/new-assessment.component.html
@@ -20,7 +20,7 @@
   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE 
   SOFTWARE. 
 -------------------------->
-<div *ngFor="let row of rows"  class="m-4" style="height:281px;">
+<div *ngFor="let row of gallerySvc.rows"  class="m-4" style="height:281px;">
     <div>
         <h3>{{row.group_Title}}</h3>
     </div>
